project/requirements.txt
Django==2.2.7
mysqlclient==1.4.5
# Python MySQL
# Use mysqlclient if you want a binary (C) driver for mysql.
# But speed advantages should be really unimportant compared to the pure-python sql client

celery==4.3.0
kombu==4.6.6

bleach==3.1.0

# botocore 1.13 does not work with python-dateutil 2.8.1 and boto3 1.10 does not work with botocore 1.12. update them
# only when botocore 1.13 will work with python-dateutil 2.8.1

boto3==1.9.251
botocore==1.12.251
cotyledon==1.7.3
cryptography==2.8
distro==1.4.0
django-cors-headers==3.1.1
django-extensions==2.2.5
django-filter==2.2.0
django-ipware==2.1.0
django-shortcuts==1.6
djangorestframework==3.10.3
gnocchiclient==7.0.5
isfreemail==0.1.5
iso8601==0.1.12
jsonfield==2.0.2
keystoneauth1==3.18.0
lxml==4.4.1
minfraud==1.7.0
oslo.config==6.11.1
oslo.context==2.23.0
oslo.log==3.44.1
oslo.messaging==10.3.0
oslo.serialization==2.29.2
paypalrestsdk==1.13.1
phonenumbers==8.10.22
pika==1.1.0
pycountry==19.8.18
pycrypto==2.6.1
pyotp==2.3.0
python-ceilometerclient==2.9.0
python-cinderclient==5.0.0
python-dateutil==2.8.1
python-designateclient==3.0.0
python-glanceclient==2.17.0
python-keystoneclient==3.22.0
python-neutronclient==6.14.0
python-novaclient==16.0.0
python-magnumclient==2.15.0
pytz==2019.3
pyvat==1.3.1
qrcode==6.1
redis==3.3.11
reportlab==3.5.32
requests==2.22.0
retrying==1.3.3
rfc3986==1.3.2
six==1.12.0
stripe==2.40.0
urllib3==1.25.7
yappi==1.0
python-stdnum==1.12
